// src/app/data-interface.ts
export interface DataInterface {
  label2: string;
  labelPhone: string;
  submitBtn: string;
  id: number;
  title: string;
  country: string;
  PersonalDetails1: string;
  labelFirstName: string;
  labelLastName: string;
  labelEmail: string;
  labelAddress: string;
  nextBtn: string;
  prevBtn: string;
  resetBtn: string;
  doneBtn: string;
  doneTxt: string;
  addressTxt: string;
  required: string;
}

export interface FormDetails {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  address: string;
  phone: string;
}